refactor(py_src): introduce ir/ package for binding-IR data classes
The classes that hold the parsed C++ shape — ClassInfo, FuncInfo,
FuncVariant, ArgInfo, ClassProp, Namespace — were scattered as
top-level modules alongside emit-side and helper code. They model
the language-neutral IR the binding generator walks, so they belong
under a `ir/` package.

  py_src/arg_info.py     -> py_src/ir/args.py
  py_src/class_prop.py   -> py_src/ir/props.py
  py_src/namespace.py    -> py_src/ir/namespaces.py
  py_src/class_info.py   -> py_src/ir/classes.py
  py_src/func_info.py    -> py_src/ir/functions.py
  py_src/func_variant.py -> py_src/ir/variants.py

ir/classes.py still imports module_generator (emit side) — that
coupling will be removed once emit is split into per-language
backends; for now it's annotated with a one-line comment.

Byte-equivalent vs lib-4.11.0/ and src-4.11.0/.
