+++ py_src/ir/namespaces.py
