+++ py_src/ir/props.py
